+++ packages/coverde_cli/CHANGELOG.md
@@ -4,6 +4,7 @@
 - **FEAT**: use Dart 3.5.0 as minimum SDK version (#114).
 - **FEAT**: add `optimize-tests` command (#115).
 - **FEAT**: add `file-coverage-log-level` option to `check` and `value` commands (#118).
+- **BREAKING FEAT**: use `base-directory` option instead of `paths-parent` option in `filter` command (#158).
 
 ## 0.2.0+2